Also renamed viewdetails to docdetails for consistency
