+++ nw/gui/__init__.py
@@ -5,6 +5,7 @@ from nw.gui.build import GuiBuildNovel
 from nw.gui.docbars import GuiDocTitleBar
 from nw.gui.docbars import GuiNoticeBar
 from nw.gui.docbars import GuiSearchBar
+from nw.gui.docdetails import GuiDocViewDetails
 from nw.gui.doceditor import GuiDocEditor
 from nw.gui.docmerge import GuiDocMerge
 from nw.gui.docsplit import GuiDocSplit
@@ -21,7 +22,6 @@ from nw.gui.sessionlog import GuiSessionLogView
 from nw.gui.statusbar import GuiMainStatus
 from nw.gui.theme import GuiIcons
 from nw.gui.theme import GuiTheme
-from nw.gui.viewdetails import GuiDocViewDetails
 
 __all__ = [
     "GuiAbout",
@@ -29,6 +29,7 @@ __all__ = [
     "GuiDocTitleBar",
     "GuiNoticeBar",
     "GuiSearchBar",
+    "GuiDocViewDetails",
     "GuiDocEditor",
     "GuiDocMerge",
     "GuiDocSplit",
@@ -45,5 +46,4 @@ __all__ = [
     "GuiMainStatus",
     "GuiIcons",
     "GuiTheme",
-    "GuiDocViewDetails",
 ]
